[Refactor]: 챌린지 기록 서비스 분리
