+++ zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/controller/challengereview/ChallengeReviewRes.java
@@ -1,4 +1,4 @@
-package org.haedal.zzansuni.controller.challengegroup.review;
+package org.haedal.zzansuni.controller.challengereview;
 
 import java.util.Map;
 import lombok.Builder;
