--- zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/domain/challengegroup/port/ChallengeStore.java
@@ -1,5 +1,0 @@
-package org.haedal.zzansuni.domain.challengegroup.port;
-
-public interface ChallengeStore {
-
-}
