+++ zzansuni-api-server/app/src/test/java/org/haedal/zzansuni/domain/challengegroup/userchallenge/UserChallengeRecordServiceTest.java
@@ -37,7 +37,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-class UserChallengeServiceTest {
+class UserChallengeRecordServiceTest {
 
     @Mock
     private UserChallengeStore userChallengeStore;
